# src/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "",
    "version": "v1",
    "title": "Flask API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
        {
        "url": "/"
        }
    ],
   "components": {
     "securitySchemes": {
       "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
      }
  },
...
        "schemas": {
            "Users": {
                "type": "object",
                "properties": {
                    "username": {
                    "type": "string"
                    },
                    "email": {
                    "type": "string"
                    }
                }
            },
...
            "Register": {
                "type": "object",
                "properties": {
                    "username": {
                    "type": "string"
                    },
                    "email": {
                    "type": "string"
                    },
                    "password": {
                    "type": "string"
                    }
                }
            },
...
           "Orders": {
                "type": "object",
                "properties": {
                    "UserID": {
                    "type": "integer"
                    },
                  "OrderNumber": {
                    "type": "string"
                    },
                  "Total": {
                    "type": "integer"
                    },
                  "Address": {
                    "type": "string"
                    },
                  "Status": {
                    "type": "string"
                    }
                }
            },
...
          "OrderNumber": {
                "type": "object",
                "properties": {
                  "OrderNumber": {
                    "type": "string"
                    }
                }
            },
          "UpdateStatus": {
                "type": "object",
                "properties": {
                  "OrderId": {
                    "type": "integer"
                    }
                }
            },
          "UpdateStatusResponse": {
                "type": "object",
                "properties": {
                  "Status": {
                    "type": "string",
                    "description": "Status message indicating the success of the operation"
                    }
                }
            },
...
          "updateUser":{
              "type": "object",
              "properties" :{
                  "Username": {
                    "type": "string"
                  },
                   "Phone": {
                    "type": "string"
                  },
                "Birth Day": {
                    "type": "string"
                    },
                  "Email": {
                    "type": "string"
                  }
              }
            },
             "updatePassword":{
              "type": "object",
              "properties" :{
                  "current_password": {
                    "type": "string"
                  },
                  "new_password": {
                    "type": "string"
                  }
              }
            },
...
        }
    },
     "paths": {
        "/api/v1/users/GetAll": {

            "get" :{
                "tags": [
                    "Users"
                ],
                "summary": "Get All Users",


                "responses": {
                "200": {
                    "description": "Success",
                    "schema": {
                    "$ref": "#/components/schemas/Users"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Users"
                    }
            }
            }
            }
        },
        "/api/v1/users/Get": {
         "get": {
        "tags": [
            "Users"
        ],
        "summary": "Get User by Id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "This endpoint retrieves user information based on the user's token information.",
        "responses": {
            "200": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Users"
                        }
                    }
                }
            },
            "404": {
                "description": "NOT FOUND",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Users"
                        }
                    }
                }
            }
        }
    }
},
       "/api/v1/users/Register": {
         "post": {
            "tags": [
                "Users"
            ],
            "summary": "Register User",
            "security": [],
            "requestBody": {
            "description": "Post New todo To Database",
            "required": true,
            "content": {
                "application/json": {
                "schema": {
                    "$ref": "#/components/schemas/Register"
                    }
                }
            }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/Users"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Users"
                    }
            }
            }
            }
       },
       "/api/v1/users/UpdateProfile": {
         "put": {
          "tags": [
            "Users"
          ],
          "summary": "Update User Informations",
           "security": [
          {
            "bearerAuth": []
          }
        ],
          "requestBody": {
            "description": "Update Existing todo On Database",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/updateUser"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/components/schemas/Users"
              }
            },
            "404": {
              "description": "NOT FOUND",
              "schema": {
                "$ref": "#/components/schemas/Users"
              }
            }
          }
        }
        },
       "/api/v1/users/ChangePassword": {
         "put": {
          "tags": [
            "Users"
          ],
          "summary": "Update User Password",
           "security": [
          {
            "bearerAuth": []
          }
        ],
          "requestBody": {
            "description": "Update Existing todo On Database",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/updatePassword"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/components/schemas/Users"
              }
            },
            "404": {
              "description": "NOT FOUND",
              "schema": {
                "$ref": "#/components/schemas/Users"
              }
            }
          }
        }
        },
...
        "/api/v1/orders/create": {
            "post": {
              "tags": [
                "Orders"
              ],
              "summary": "Create Order",
              "security": [
          {
            "bearerAuth": []
          }
        ],
              "requestBody": {
              "description": "Post New todo To Database",
              "required": true,
              "content": {
                  "application/json": {
                  "schema": {
                      "$ref": "#/components/schemas/Orders"
                     }
                 }
              }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/Orders"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Orders"
                    }
            }
            }
            }
       },
        "/api/v1/orders/GetAll": {
          "get" :{
                "tags": [
                    "Orders"
                ],
                "summary": "Get All ",
            "security": [
          {
            "bearerAuth": []
          }
        ],
                "responses": {
                "200": {
                    "description": "Success",
                    "schema": {
                    "$ref": "#/components/schemas/Orders"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Orders"
                    }
            }
            }
            }
       },
        "/api/v1/orders/updateStatus": {
            "post": {
              "tags": [
                "Orders"
              ],
              "summary": "Update Status Order",
              "requestBody": {
              "description": "Post New todo To Database",
              "required": true,
              "content": {
                  "application/json": {
                  "schema": {
                      "$ref": "#/components/schemas/UpdateStatus"
                     }
                 }
              }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/UpdateStatusResponse"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/UpdateStatusResponse"
                    }
            }
            }
            }
       },
        "/api/v1/orders/getOrderIdByOrderNumber": {
          "get": {
            "tags": [
              "Orders"
            ],
            "summary": "Get OrderID by OrderNumber",
            "security": [
          {
            "bearerAuth": []
          }
        ],
       "parameters": [
          {
            "name": "orderNumber",
            "in": "query",
            "description": "ID of user to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
            "responses": {
              "200": {
                "description": "OK",
                "schema": {
                  "$ref": "#/components/schemas/OrderNumber"
                }
              },
              "404": {
                "description": "NOT FOUND",
                "schema": {
                  "$ref": "#/components/schemas/OrderNumber"
                }
              }
            }
          }
        },
...
        }
  }